Backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy>=2.0
asyncpg
psycopg2-binary

pydantic>=2.0
pydantic-settings

python-jose
passlib[bcrypt]

python-dotenv

alembic
